refactor(linter): split word helpers by layer
