+++ crates/shuck-linter/src/rules/correctness/glob_in_test_comparison.rs
@@ -1,4 +1,6 @@
-use crate::{Checker, Rule, SimpleTestShape, SimpleTestSyntax, Violation, static_word_text};
+use shuck_ast::static_word_text;
+
+use crate::{Checker, Rule, SimpleTestShape, SimpleTestSyntax, Violation};
 
 pub struct GlobInTestComparison;
