+++ crates/shuck-linter/src/rules/correctness/syntax.rs
@@ -1,6 +1,4 @@
-use shuck_ast::{Assignment, SimpleCommand, Word};
-
-pub(crate) use crate::static_word_text;
+use shuck_ast::{Assignment, SimpleCommand, Word, static_word_text};
 
 fn simple_command_name(command: &SimpleCommand, source: &str) -> Option<String> {
     static_word_text(&command.name, source).map(|text| text.into_owned())
