+++ crates/shuck-linter/src/rules/style/syntax.rs
@@ -1,6 +1,4 @@
-use shuck_ast::Command;
-
-pub(crate) use crate::static_word_text;
+use shuck_ast::{Command, static_word_text};
 
 pub fn is_simple_command_named(command: &Command, source: &str, name: &str) -> bool {
     match command {
