+++ crates/shuck-linter/src/rules/style/unquoted_path_in_mkdir.rs
@@ -1,6 +1,6 @@
-use crate::{
-    Checker, CommandFact, ExpansionContext, Rule, Violation, WordFactContext, static_word_text,
-};
+use shuck_ast::static_word_text;
+
+use crate::{Checker, CommandFact, ExpansionContext, Rule, Violation, WordFactContext};
 
 pub struct UnquotedPathInMkdir;
