+++ crates/shuck-linter/src/rules/style/unquoted_variable_in_test.rs
@@ -1,6 +1,7 @@
+use shuck_ast::static_word_text;
+
 use crate::{
     Checker, ExpansionContext, Rule, SimpleTestShape, SimpleTestSyntax, Violation, WordFactContext,
-    static_word_text,
 };
 
 pub struct UnquotedVariableInTest;
